tests: qdma: fix the driver mode
The driver MODE string should be: device-number:0:mode.  Unfortunately the
driver won't complain if the device doesn't exist, and the MODE string were
applicable to the first device.

Fix that by making a configurable DEVICE variable.  It has "02" as a default
value which is compatible with the versal-qdma-cpm4 demo, and can be altered
by setting it prior to call the test script.

Also a sanity check is done, by checking if the sysfs qmax variable is present,
to be sure that the device is the correct one.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>