qdma: fix aggregated irq ring index
For the CPM5 in aggregated interrupt mode, the IRQ ring is selected by the vec
field in the software context and not by the queue id as it was implemented.

This fixes that wrong behavior.  It also removes the mention of 'qid' when
accessing the irq ring to avoid any confusion.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>